vent function
Idea is to have a 2 stage vent one to prime the ecig and then a fan to vent it out. I have never figured out how to use millias as a delay so inserted a 600ms delay between the 2 stages. Also put canceling code in areas so if switches get turn off at the wrong time the ecig will not stay on when song is played and when any thing is shut down. lastly put in a servo for pin 13 pwm so can be wired to have cores eject out like in the video game or like the phoenix pack. my light numbers are different from useing led kits i had here instead of the parts suggested.